Practica XD
en fin la GTC CLNA

+++ MP07 Desarrolo web en entorno servidor/segundoProyecto/database/migrations/2022_11_04_163105_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string("name");
-            $table->integer("price");
+            $table->float("price");
             $table->integer("stock");
             $table->string("description");
             $table->unsignedBigInteger("idCategory");
